# Fix RTT calculation for accurate timing measurement