## Summary
- Fixed RTT calculation that was showing unrealistically short times for local hosts and incorrect timing for remote hosts
- Implemented timestamp storage directly in ICMP packet payload for precise round-trip time measurement
- Corrected socket configuration to prevent timing interference

## Changes Made
- **Accurate timestamp capture**: Store send timestamp in ICMP packet payload instead of measuring from receive loop start
- **Proper timing validation**: Only capture receive timestamp after validating the packet is our ICMP reply
- **Socket configuration fixes**: 
  - Switched from non-blocking to blocking mode to prevent timing artifacts
  - Reduced timeout from 1s to 100ms to minimize interference
- **Packet validation improvements**: Remove strict sequence validation to handle out-of-order responses

## Test Results
**Before fix:**
- Localhost: Impossible <1ms times
- Remote hosts (8.8.8.8): Unrealistic <1ms times

**After fix:**
- Localhost: Proper microsecond precision (0.045-0.117ms)
- Google DNS (8.8.8.8): Realistic network latencies (8-12ms)
- AliExpress: Shows network timing (varies by network conditions)

## Test plan
- [x] Test localhost ping timing accuracy
- [x] Test remote host ping (Google DNS) for realistic RTT values
- [x] Verify timing precision matches standard ping behavior
- [x] Confirm no regression in packet loss statistics
- [x] Test with verbose output for debugging

🤖 Generated with [Claude Code](https://claude.ai/code)